chore: rename dashboard EscrowDashboard to RoleEscrowDashboard
Resolves naming collision with escrow/EscrowDashboard and updates all
import paths for the role-based dashboard component and shared types.

Closes #359
